[widget] set the focus in rdp_input_mouse() for avoid double refresh
When WidgetComposite propagate mouse event, focus is also managed.
When a button — or other widget — take the focus, a refresh is in focus()
then an other is in WidgetButton::rdp_input_mouse()

Now, has_focus is false when rdp_input_mouse() is called.
If after has_focus is still false, focus() is called.
Otherwise, it means that the focus is managed by the widget.