[session] stop capture before open the close box with cluster and session inactivity (fix WAB-15202)
When a cluster is configured, `has_more_target` is set to true if not all targets
have been tested, so that logging and recording are not stopped when trying another
target.

This state remains even when the session starts normally, but at that point,
`next_backend_module()` must stop capture.

The fix is to ignore `has_more_target` if a module other than the current one
is requested (RDP in this context).